weaviate-sentiment-analysis/requirements.txt
pandas==1.2.4
weaviate-client==3.0.0
